requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Testing==0.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
py-healthcheck==1.7.0
pymongo==3.6.1
six==1.11.0
virtualenv==15.1.0
Werkzeug==0.14.1
